PVS-Studio V2009: explicit type conversion from enum type to integer type

+++ lre.h
@@ -309,7 +309,7 @@ lre_tag_t lrex_tag_by_nbytes_negative(int nbytes) {
  */
 lre_decl
 int lrex_nbytes_by_tag_positive(lre_tag_t numeric_tag) {
-	return numeric_tag - (LRE_TAG_NUMBER_POSITIVE_1 - 1);
+	return (int) numeric_tag - ((int) LRE_TAG_NUMBER_POSITIVE_1 - 1);
 }
 
 
@@ -319,7 +319,7 @@ int lrex_nbytes_by_tag_positive(lre_tag_t numeric_tag) {
  */
 lre_decl
 int lrex_nbytes_by_tag_negative(lre_tag_t numeric_tag) {
-	return (LRE_TAG_NUMBER_NEGATIVE_1 + 1) - numeric_tag;
+	return ((int) LRE_TAG_NUMBER_NEGATIVE_1 + 1) - (int) numeric_tag;
 }
 
 
@@ -628,7 +628,7 @@ int lre_pack_str(lre_buffer_t *buf, const uint8_t *src, size_t len, lre_mod_t mo
 		}
 		
 		lrex_write_str (&dst, src, len, 0);
-		lrex_write_char(&dst, mod);
+		lrex_write_char(&dst, (int) mod);
 		lrex_write_char(&dst, LRE_SEP_POSITIVE);
 		lre_buffer_set_size_distance(buf, dst);
 		return LRE_OK;
@@ -648,14 +648,14 @@ int lre_pack_int(lre_buffer_t *buf, int64_t value, lre_error_t *error) {
 			uint64_t uvalue = lrex_negate_negative(value);
 			uint8_t  nbytes = lrex_count_nbytes(uvalue);
 			
-			lrex_write_char   (&dst, lrex_tag_by_nbytes_negative(nbytes));
+			lrex_write_char   (&dst, (int) lrex_tag_by_nbytes_negative(nbytes));
 			lrex_write_uint64n(&dst, ~uvalue, nbytes);
 			lrex_write_char   (&dst, LRE_SEP_NEGATIVE);
 		}
 		else {
 			uint8_t nbytes = lrex_count_nbytes(value);
 			
-			lrex_write_char   (&dst, lrex_tag_by_nbytes_positive(nbytes));
+			lrex_write_char   (&dst, (int) lrex_tag_by_nbytes_positive(nbytes));
 			lrex_write_uint64n(&dst, value, nbytes);
 			lrex_write_char   (&dst, LRE_SEP_POSITIVE);
 		}
@@ -704,7 +704,7 @@ int lre_pack_float(lre_buffer_t *buf, double value, lre_error_t *error) {
 			/* Decimal inversion */
 			fraction = lrex_max10[15] - fraction;
 			
-			lrex_write_char   (&dst, lrex_tag_by_nbytes_negative(nbytes));
+			lrex_write_char   (&dst, (int) lrex_tag_by_nbytes_negative(nbytes));
 			lrex_write_uint64n(&dst, ~integral, nbytes);
 			lrex_write_decimal(&dst, fraction, 15);
 			lrex_write_char   (&dst, LRE_SEP_NEGATIVE);
@@ -714,7 +714,7 @@ int lre_pack_float(lre_buffer_t *buf, double value, lre_error_t *error) {
 			uint64_t fraction = (value - integral) * 1e15 + 0.5;
 			uint8_t  nbytes   = lrex_count_nbytes(integral);
 			
-			lrex_write_char   (&dst, lrex_tag_by_nbytes_positive(nbytes));
+			lrex_write_char   (&dst, (int) lrex_tag_by_nbytes_positive(nbytes));
 			lrex_write_uint64n(&dst, integral, nbytes);
 			lrex_write_decimal(&dst, fraction, 15);
 			lrex_write_char   (&dst, LRE_SEP_POSITIVE);
